@@ -4,14 +4,14 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^0.1.2
-  google_maps_flutter: ^0.5.21+6
+  google_maps_flutter: ^2.0.6
 
 dev_dependencies:
   flutter_test:
