@@ -5,14 +5,14 @@ author: Shakir Samoon <shakirsamoon@gmail.com>
 homepage: https://github.com/Shark01/google_map_polyline
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  google_maps_flutter: ^0.5.21+6
-  dio: ^3.0.2
+  google_maps_flutter: ^2.0.6
+  dio: ^4.0.0
 
 dev_dependencies:
   flutter_test:
