@@ -8,17 +8,17 @@ import 'package:google_map_polyline/src/polyline_utils.dart';
 /// Gives polyline coordinates to set polylines in Google Map.
 class GoogleMapPolyline {
   String apiKey;
-  GoogleMapPolyline({@required this.apiKey});
+  GoogleMapPolyline({required this.apiKey});
 
-  PolylineUtils _utils;
-  PolylineRequestData _data;
+  late PolylineUtils _utils;
+  PolylineRequestData? _data;
 
   /// Get coordinates using Location Coordinates
   /// Example : LatLng(40.677939, -73.941755)
-  Future<List<LatLng>> getCoordinatesWithLocation({
-    @required LatLng origin,
-    @required LatLng destination,
-    @required RouteMode mode,
+  Future<List<LatLng>?> getCoordinatesWithLocation({
+    required LatLng origin,
+    required LatLng destination,
+    required RouteMode mode,
   }) async {
     _data = new PolylineRequestData(
         originLoc: origin,
@@ -33,9 +33,9 @@ class GoogleMapPolyline {
   }
 
   Future<List<RoutesWithSummary>> getCoordinatesWithLocationAndAlternatives({
-    @required LatLng origin,
-    @required LatLng destination,
-    @required RouteMode mode,
+    required LatLng origin,
+    required LatLng destination,
+    required RouteMode mode,
   }) async {
     _data = new PolylineRequestData(
         originLoc: origin,
@@ -51,10 +51,10 @@ class GoogleMapPolyline {
 
   /// Get coordinates using Location Coordinates
   /// Example : '55 Kingston Ave, Brooklyn, NY 11213, USA'
-  Future<List<LatLng>> getPolylineCoordinatesWithAddress({
-    @required String origin,
-    @required String destination,
-    @required RouteMode mode,
+  Future<List<LatLng>?> getPolylineCoordinatesWithAddress({
+    required String origin,
+    required String destination,
+    required RouteMode mode,
   }) async {
     _data = new PolylineRequestData(
         originText: origin,
